// components/business/forms/custom/accident-claim-form.d.ts
export interface AccidentClaimFormData {
  formUrl: string;

  // Patient Personal Information
  full_name: string;
  email: string;
  country: string;
  state: string;
  primary_contact: string;
  other_contact_name: string;
  other_contact_phone: string;

  // Accident Information
  accident_date: Date | null | string;
  accident_place: string;
  accident_type: string;
  sub_accident_type: string;

  // Motor Vehicle Accident Details
  mva_type: string;
  mva_location: string;
  vehicle_details: Array<{ insuranceName: string; policyNumber: string }>;
  selected_vehicle: string;
  mva_description: string;
  mva_medical_info: {
    assistanceType: string;
    diagnosis: string;
    treatment: string;
    primaryCareProvider: string;
  };
  mva_costs: {
    totalCost: string;
    policyLimits: string;
    assistanceStatus: string;
    medicalProviderCosts: string;
    repatriationCosts: string;
    otherCosts: string;
  };
  mva_third_party_info: {
    insuranceCompany: string;
    claimReferenceNumber: string;
    adjusterName: string;
    adjusterContactDetails: string;
    ownerBusinessName: string;
    ownerReferenceNumber: string;
    ownerPhoneNumber: string;
    coInsured: string;
    otherPartyInfo: string;
  };
  mva_attorney_info: {
    lawFirmName: string;
    attorneyName: string;
    attorneyPhone: string;
  };

  // Slip and Fall Details
  slip_description: string;
  slip_accident_type: string;
  negligence_description: string;
  witness_info: {
    name: string;
    email: string;
    phone: string;
  };
  slip_medical_info: {
    assistanceType: string;
    diagnosis: string;
    treatment: string;
    primaryCareProvider: string;
  };
  slip_costs: {
    totalCost: string;
    policyLimits: string;
    assistanceStatus: string;
    medicalCost: string;
    repatriationCosts: string;
    otherCosts: string;
  };
  slip_third_party_info: {
    insuranceCompany: string;
    claimReferenceNumber: string;
    adjusterName: string;
    adjusterContactDetails: string;
    ownerBusinessName: string;
    ownerReferenceNumber: string;
    ownerPhoneNumber: string;
    coInsuredName: string;
    otherPartyInfo: string;
  };
  slip_attorney_info: {
    lawFirmName: string;
    attorneyName: string;
    attorneyPhone: string;
  };

  // File Uploads
  documentFiles: FileList | null;
  mvaUploadDocumentation: FileList | null;
  mvaRepatriationBills: FileList | null;
  mvaOtherFiles: FileList | null;
  mvaInsuranceDocs: FileList | null;
  mvaBusinessDocs: FileList | null;
  mvaCoInsuredDocs: FileList | null;
  mvaAttorneyDocs: FileList | null;

  slipAccidentReports: FileList | null;
  slipPhotos: FileList | null;
  slipMedicalDocs: FileList | null;
  slipMedicalBills: FileList | null;
  slipRepatriationBills: FileList | null;
  slipThirdPartyDocs: FileList | null;
  slipBusinessDocs: FileList | null;
  slipCoInsuredDocs: FileList | null;
}